Update key character map
We update the generation script to be able to use the 102ND key, and
remap all keys to QWERTY layout in order to use the BÉPO character map
consistently, even on AZERTY and other keyboards.

The generated file now has a license, and it's the same as this
project's.